# docs/swagger.yaml
basePath: /api
definitions:
...
  model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.HttpError'
    type: object
...
  model.Friend:
    properties:
      id:
        type: string
      image:
        type: string
      isOnline:
        type: boolean
      username:
        type: string
    type: object
  model.FriendRequest:
    properties:
      id:
        type: string
      image:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.RequestType'
        description: '1: Incoming, 0: Outgoing'
        enum:
        - 0
        - 1
      username:
        type: string
    type: object
...
  model.HttpError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
  model.RequestType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Outgoing
    - Incoming
  model.Success:
    properties:
      success:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact:
    email: khoingoc456@gmail.com
    name: KhoiPN
  description: MurmurChatapp REST API Specs. This service uses sessions for authentication
  title: MurmurChatapp API
  version: "1.0"
paths:
...
  /account/{memberId}/friend:
    delete:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove Friend
      tags:
      - Friends
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Send Friend Request
      tags:
      - Friends
  /account/{memberId}/friend/accept:
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Accept Friend's Request
      tags:
      - Friends
  /account/{memberId}/friend/cancel:
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cancel Friend's Request
      tags:
      - Friends
...
  /account/me/friends:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Friend'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Current User's Friends
      tags:
      - Friends
  /account/me/pending:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.FriendRequest'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Current User's Friend Requests
      tags:
      - Friends
...
swagger: "2.0"